requirements.txt
aiohttp
python-dotenv
bs4